Rename mkosi.distributions to mkosi.distribution
We have mkosi.installer, not mkosi.installers, so to keep things
consistent let's use mkosi.distribution instead of
mkosi.distributions.
