+++ mkosi/distribution/azure.py
@@ -4,7 +4,7 @@ from collections.abc import Iterable
 
 from mkosi.config import Architecture
 from mkosi.context import Context
-from mkosi.distributions import (
+from mkosi.distribution import (
     Distribution,
     fedora,
     join_mirror,
