+++ mkosi/distribution/centos.py
@@ -5,7 +5,7 @@ from collections.abc import Iterable
 from mkosi.config import Architecture, Config
 from mkosi.context import Context
 from mkosi.curl import curl
-from mkosi.distributions import (
+from mkosi.distribution import (
     Distribution,
     DistributionInstaller,
     PackageType,
