+++ mkosi/distribution/rhel_ubi.py
@@ -3,7 +3,7 @@
 from collections.abc import Iterable
 
 from mkosi.context import Context
-from mkosi.distributions import Distribution, centos, join_mirror
+from mkosi.distribution import Distribution, centos, join_mirror
 from mkosi.installer.rpm import RpmRepository, find_rpm_gpgkey
 from mkosi.log import die
